Initial commit
The initial commit contains a LICENSE (MPL 2.0) file, git files, a
Makefile[1] for building etc., and an EditorConfig[2] file.

On top of that, it contains an initial version of the wordrow
application structured in accordance with [3].

The wordrow application aims to be a simple CLI tool that can, somewhat
intelligently, replace words by other words in text files. An example
use case would be to replace all American English words by British
English words in a LaTeX project. Another example use case would be to
automatically fix common mistakes in MarkDown formatted documentation.

--
1. https://www.gnu.org/software/make/manual/make.html
2. https://editorconfig.org/
3. https://github.com/golang-standards/project-layout