Add open source files an Code Climate integration
Add a README file, CHANGELOG file, and CODE_OF_CONDUCT file to the
repository.

Also added Code Climate integration (maintainability and coverage),
including badges to the README. The README also has a badge for the
GitHub Action CI. In addition, the README contains same basic text about
the project.

The CHANGELOG is currently empty and follows the format described by the
Keep a Changelog project [1].

The CODE_OF_CONDUCT is adapted from the Contributor Covenant Code of
Conduct version 2.0 [2].

--
1. https://keepachangelog.com/en/1.0.0/
2. https://www.contributor-covenant.org/version/2/0/code_of_conduct.html