Implement glob support for input file
Add a new function called `ResolveGlobs` to the fs package which can be
used to resolve globs into file paths, using the Glob function of the
filepath package [1] of Go.

This function it tested for a scenario where:
- there are no arguments
- there are only arguments that are not globs
- there are arguments that are globs
- there is an argument that is an invalid glob

Finally, this function is used to resolve globs of input files and
mapping files.

--
1. https://golang.org/pkg/path/filepath/#Glob