Add support to define prefix and suffix matching in mapping files
Which replaces the old behaviour of never matching words with a prefix 
and always matching words with a suffix.

In the current implementation, the prefix and suffix token (`-`) should 
only be used in the "from" value. In the "to" value it will not be 
detected and simply be used as part of the replacement.