Simplify Mapping type
First, creating a Mapping is now (again) as easy as just giving it the 
"from" and "to" value, instead of encoding this in separate types 
(`fromValue` and `toValue` resp.). To do this, the logic that was 
previously distruted over a couple of types, is now all in (private) 
methods on the Mapping type.

Second, the Match type now includes a field name "Replacement" that 
contains the value that would be returned by GetReplacement if you 
called it with the "Prefix" and "Suffix" values of the Match. As such, 
the "Prefix" and "Suffix" fields have been removed from the Match type.