Add linting of MarkDown files
Adds a Make command, "make lint", that can be used to lint MarkDown
files (e.g. documentation). To do this, markdownlint-cli [1] is used. As
this is a NodeJS tool available through NPM, I utlised npx. npx is a
tool from npm that allows you to run it's package on-demand by
downloading them only once they're necessary. Note that this does
require that NodeJS is installed on your system!

markdownlint-cli is currently used to lint all files in the project with
the md-extension. This include the CHANGELOG, CODE_OF_CONDUCT, and
README, as well as the file(s) in the ./docs folder. The configuration
that is used when running markdownlint can be found in the new file
".markdownlintrc.yml" (formatted in YAML).

The chosen configuration for markdownlint is incompatible with the style
of the CODE_OF_CONDUCT. Therefore, it has been changed to no longer
conflict with the configuration. This entails:

- Change the bullets used for unordered list from "*" to "-".
- Replace instance of bare URLs with MarkDown links.

Moreover, the GitHub Actions workflow file has been updated to perform
this lint automatically on every push and Pull Request.

--
1. https://www.npmjs.com/package/markdownlint-cli