Account for hyphens and other non-space characters that maintain caps
This improves the functionality that maintains capitalization for multi-
word values. Now, if the to value is split by e.g. a hyphen, instead of
a space, capitalization of the second word is still maintained.

Moreover, this fixes a potential bug where a multi-word value contains
multiple spaces between its words. The previous implementation assumed
a single space. The new implementation takes the division symbol(s)
found in the mapping.