Add basic documentation on defining mappings in (CSV) files
This adds basic documentation on how to write a mapping file. The
examples are given in the CSV format, but are applicable for MarkDown
just as well. The document covers 1) defining a simple mapping of words
or phrases, 2) that leading and trailing whitespace is ignored, but 3)
that whitespace in a phrase is never ignored, and 4) that the order of
mappings in a file matters.

In addition, it adds a test for the `replacer` package that verifies all
of the documented behaviour is actually true. The facts about whitespace
within a phrase was not test yet.

Lastly, this fixes the `[mapping file]` link in the CLI documentation
and updates the markdownlint configuration to disallow using emphasis as
headings (the format I envisioned that made me allow it was not used in
the end).