Add documentation on the available mapping formats
This commit adds a third documentation file that explains the different
mapping formats that are supported. Currently, these are CSV and
MarkDown. The document gives 1) some basic information on how the file
is interpreted, 2) an example, 3) reasons why the file may be considered
invalid, and 4) a list of file extensions that are considered to be CSV
or MarkDown.

Also, this commit addresses the broken relative links in the
documentation files.