Update logger to take into account a LogLevel
The log level is a new type introduced in this commit that is defined in 
internal/log_level.go. This log level is used in two new funtion in the 
existin logger package.

First there now is one internal function that does all the actual 
printing. It will decide of something should be logged based on the 
assigned LogLevel, and also format the message accordingly (i.e. add 
e.g. "[Error] " in front of it).

Second, there is a function called `SetLogLevel` which can be used from 
outside the package to set the maximum log level that should be 
outputted in this program run.

Note that the log level is program wide, so if it is set in one place it 
applies everywhere. Given the currently estimated scope of this project, 
that should be sufficient.

The (new) LogLevel enum is fully documented and tested.