Create documentation on tool usage (#24)
* Add basic documentation on using the wordrow CLI

This adds a MarkDown formatted `./docs`umentation file that explains, 
briefly, how the CLI of wordrow can be used to replace words in 
plaintext files by other words.


* Add linting of MarkDown files

Adds a Make command, "make lint", that can be used to lint MarkDown
files (e.g. documentation). To do this, markdownlint-cli [1] is used. As
this is a NodeJS tool available through NPM, I utlised npx. npx is a
tool from npm that allows you to run it's package on-demand by
downloading them only once they're necessary. Note that this does
require that NodeJS is installed on your system!

markdownlint-cli is currently used to lint all files in the project with
the md-extension. This include the CHANGELOG, CODE_OF_CONDUCT, and
README, as well as the file(s) in the ./docs folder. The configuration
that is used when running markdownlint can be found in the new file
".markdownlintrc.yml" (formatted in YAML).

The chosen configuration for markdownlint is incompatible with the style
of the CODE_OF_CONDUCT. Therefore, it has been changed to no longer
conflict with the configuration. This entails:

- Change the bullets used for unordered list from "*" to "-".
- Replace instance of bare URLs with MarkDown links.

Moreover, the GitHub Actions workflow file has been updated to perform
this lint automatically on every push and Pull Request.

--
1. https://www.npmjs.com/package/markdownlint-cli


* Add basic documentation on defining mappings in (CSV) files

This adds basic documentation on how to write a mapping file. The
examples are given in the CSV format, but are applicable for MarkDown
just as well. The document covers 1) defining a simple mapping of words
or phrases, 2) that leading and trailing whitespace is ignored, but 3)
that whitespace in a phrase is never ignored, and 4) that the order of
mappings in a file matters.

In addition, it adds a test for the `replacer` package that verifies all
of the documented behaviour is actually true. The facts about whitespace
within a phrase was not test yet.

Lastly, this fixes the `[mapping file]` link in the CLI documentation
and updates the markdownlint configuration to disallow using emphasis as
headings (the format I envisioned that made me allow it was not used in
the end).


* Add documentation on the available mapping formats

This commit adds a third documentation file that explains the different
mapping formats that are supported. Currently, these are CSV and
MarkDown. The document gives 1) some basic information on how the file
is interpreted, 2) an example, 3) reasons why the file may be considered
invalid, and 4) a list of file extensions that are considered to be CSV
or MarkDown.

Also, this commit addresses the broken relative links in the
documentation files.