Document how prefix and suffix notation can be used
Updates the documentation on mapping files to include a section on
defining that include a prefix, suffix, or both. It covers the intuitive
usage where the prefix/suffix is maintained when mapping. It also covers
that 1) it is also considered a match if there is no prefix or suffix
for the word, 2) you can match both prefixes and suffixes in one
mapping, 3) that dashes in the middle of words don't mean anything
special, and finally 4) that omitting the dash in the "to" value changes
wordrow's behaviour.