Add debug logging and implement --silent & --verbose (#21)
* Add Debug(f) method to logger

In preparation for the --verbose option, which should show debug logs.

In addition, this commit includes on example use case of debug logging 
and tests for the new methods.

* Update logger to take into account a LogLevel

The log level is a new type introduced in this commit that is defined in 
internal/log_level.go. This log level is used in two new funtion in the 
existin logger package.

First there now is one internal function that does all the actual 
printing. It will decide of something should be logged based on the 
assigned LogLevel, and also format the message accordingly (i.e. add 
e.g. "[Error] " in front of it).

Second, there is a function called `SetLogLevel` which can be used from 
outside the package to set the maximum log level that should be 
outputted in this program run.

Note that the log level is program wide, so if it is set in one place it 
applies everywhere. Given the currently estimated scope of this project, 
that should be sufficient.

The (new) LogLevel enum is fully documented and tested.

* Implement the --silent and --verbose flag

Simply by setting the log level based on which flag the program saw.

Currently, the --silent flag takes precedent. I.e., if both --silent and 
--verbose are provided as argument, then --silent is used; regardless of 
which was specified first/last.

* Update CHANGELOG

Add the two features implemented in this branch to the CHANGELOG.
Namely: an implementation that supports the `--silent` flag and the
`-verbose` flag.

* Add now working --silent and --verbose flags to documentation

Update the CLI documentation to include usage of the newly implemented 
--silent and --verbose flags.

In addition, this also adds the somewhat related --dry-run flag to the 
CLI documentation.