Refactor Mapping.Match and improve it's documentation
First off, this adds some missing documentation and improves some of the
existing documentation in .../internal/wordmap/mapping.go

Besides that, this refactors Mapping.Match, using an internal function
to find matches, and private methods to do the filtering.

Lastly, this fixes an issue where a Match's `Prefix` and `Suffix` were
not assigned.