Add support for prefix and suffix matching (#22)
* Add support to define prefix and suffix matching in mapping files

Which replaces the old behaviour of never matching words with a prefix 
and always matching words with a suffix.

In the current implementation, the prefix and suffix token (`-`) should 
only be used in the "from" value. In the "to" value it will not be 
detected and simply be used as part of the replacement.

* Add support to maintain or omit a prefix or suffix

Which changes the old behaviour where the prefix or suffix was always
maintained. There are now separate test cases for the case where a
prefix or suffix should be maintained and for the case where they should
not.

* Simplify Mapping type

First, creating a Mapping is now (again) as easy as just giving it the 
"from" and "to" value, instead of encoding this in separate types 
(`fromValue` and `toValue` resp.). To do this, the logic that was 
previously distruted over a couple of types, is now all in (private) 
methods on the Mapping type.

Second, the Match type now includes a field name "Replacement" that 
contains the value that would be returned by GetReplacement if you 
called it with the "Prefix" and "Suffix" values of the Match. As such, 
the "Prefix" and "Suffix" fields have been removed from the Match type.

* Add prefix/suffix matching to the CHANGELOG

* Document how prefix and suffix notation can be used

Updates the documentation on mapping files to include a section on
defining that include a prefix, suffix, or both. It covers the intuitive
usage where the prefix/suffix is maintained when mapping. It also covers
that 1) it is also considered a match if there is no prefix or suffix
for the word, 2) you can match both prefixes and suffixes in one
mapping, 3) that dashes in the middle of words don't mean anything
special, and finally 4) that omitting the dash in the "to" value changes
wordrow's behaviour.

* Minor unrelated changes

- Update the EditorConfig file to allow longer lines in test files.