Use 'filepathx' to resolve double star globs
This fixes a "bug" in the existing implementation where the double-star
(`**`) operator is not supported by the glob functionality for input
files and mapping files.

This is because Go's standard implementation to resolve globs does not
support the double star operator.

Following [1], I used a tiny Gomodule dependency to add double star
support in globs by changing just one line.

--
1. https://stackoverflow.com/a/40181899