Add dependency to go.mod and add go.sum
This adds the filepathx dependency (used for double star support in
globs) to the go.mod file and adds a go.sum file to ensure the correct
version of filepathx is used.

Furthermore, this adds go.sum to .gitattributes with the flag to not
show diffs. This is because it is machine generated and not intended
for humans to be read.