Version bump
This increases the version number from 'v0.1.0-beta' to 'v0.2.0-beta'.
The CHANGELOG is updated accordingly, so is the output of `--version`.

Also, based on [1] this commit changes the versioning system to semantic
versioning [2] and adds release notes for 'v0.1.0-beta' (i.e. the
version in the initial commit: 7d7b99a217e1a066c9630209b036418ec7893aa1)

--
1. https://blog.golang.org/publishing-go-modules
2. https://semver.org/spec/v2.0.0.html