Proof of concept of CLI specified mappings
This adds a proof of concept of specifying a mapping from the CLI to
replace instances of one word by another in files. Currently, the
placeholder option --tmp (and alias -t) allow you to specify a comma-
separated value that maps one word to another. If spaces are required,
the option value must be surrounded by quotations marks as otherwise the
program won't receive it as a single value.

NOTE: this is not a final implementation. Significant changes are
required before this can be merged into the master branch.