Rename --map to --map-file and -m to -M
Rename --map to --map-file and -m to -M in preparation for adding the
--map and -m flag which can be used to specify a single mapping from the
CLI.

It was decided that the current option (to specify a file) can be "more
cumbersome" to use as it is less likely to be used in a hurry. Mapping
files are more likely to be used in automated environments, or to be
reused often from the CLI (where you can use your CLI history). The
planned --map/-m option may be used "to quickly replace a word or two in
some documents." Therefore, it makes sense the latter is easier/quicker
to use.

Also update th CHANGELOG to include this change.

BREAKING CHANGE: renaming the option to specify a map file means that
automated usage of wordrow needs to be updated to use the new option
name or alias.