Refactor CLI package (#30)
* Rename option variables that are flags to use flag suffix

* Update printing of usage message

First, this adds the --invert and --verbose flags to the usage message.
Second, this splits the usage message into "flags" and "options". And,
third, this refactors the `printOption` function so that the explanation
message can be written using multiline strings without affecting the
printed message.

* Convert argContext to string using formatted strings

Refactor the `argContext.String()` method to use `fmt.Sprintf` to create 
the string versions of the configFile context and mapFile context using 
the option and optionAlias constants.

* Rename --map to --map-file and -m to -M

Rename --map to --map-file and -m to -M in preparation for adding the
--map and -m flag which can be used to specify a single mapping from the
CLI.

It was decided that the current option (to specify a file) can be "more
cumbersome" to use as it is less likely to be used in a hurry. Mapping
files are more likely to be used in automated environments, or to be
reused often from the CLI (where you can use your CLI history). The
planned --map/-m option may be used "to quickly replace a word or two in
some documents." Therefore, it makes sense the latter is easier/quicker
to use.

Also update the CHANGELOG to include this change.

BREAKING CHANGE: renaming the option to specify a map file means that
automated usage of wordrow needs to be updated to use the new option
name or alias.

* Update comments for flag aliases

* Update documentation regarding changes to program options