Refactor CLI parse defaults testing
Update the CLI parse test suite to simplify the process of testing the 
default values of arguments not being tested.

Before, each default was a separate function call in the test in 
question. Now there is a single function to call to test the default 
values of options not being tested.