Add CLI specified mappings (#31)
* Proof of concept of CLI specified mappings

* Add debug logging for CLI specified mappings

Add debug logging for the CLI specified mappings and fix the error 
produced upon a faulty CLI specified mapping.

* Update CHANGELOG

* Set '--map' and '-m' for CLI based mappings

* Test new CLI argument

Add tests for the newly introduced -m/--map CLI option and the 
corresponding `argContext`.

* Refactor CLI parse defaults testing

Update the CLI parse test suite to simplify the process of testing the 
default values of arguments not being tested.

Before, each default was a separate function call in the test in 
question. Now there is a single function to call to test the default 
values of options not being tested.

* Create bad weather tests for --map/-m option

Add bad weather tests similar to those for --map-file and --config-file for the new --map/-m option. This verifies the following two things:

1. If no value is provided for the --map/-m option the program won't run
2. If another option, instead of a value, is given after the --map/-m option the program won't run

* Update documentation to include the new --map option