Refactor `maintainCapitalization` to remove the ened for utils.Shortest
This refactors the `maintainCapitalization` function so that the 
`Shortest` function of the new utils package (now removed again) is no 
longer needed.

It also moves the Regular Expression used in `maintainCapitalization` 
outside the function definition.

Lastly, it simplifies the usage of the previously named `isUpperChar` 
function. It now accepts a string and returns whether the first 
character is uppercase. This was it's only use case.