Set up golint (#36)
* Add golint dependency and run with `make lint`

Add golint [1] as a dependency for the project and update the make command to run the golint on the project's source files. In addition there are now three lint targets: 1) run all linting, 2) run Go linting, 3) run MarkDown linting.

Also update the GitHub Action's Lint job to only lint MarkDown files.

--
1. https://github.com/golang/lint

* Fix existing golint issues

Most existing issues are problems with documentation for exported 
functions, which should start with the function's name. Another commen 
problem in the current source code is the use of `else` after a return 
in the `if`.

* Run golint in GitHub Actions

This updates the go.yml Workflow file to run golint using `make lint-go`
in GitHub Actions. To do this, the "lint" job has been updated:

- Now set ups both Go 1.13 and NodeJS 12.x
- There is now a comment why NodeJS is set up (as this is non-trivial)
- The job now has a step that properly checks out the code
- The job now has a step that installs golint.
- The job now has a step that runs golint from the Makefile.

On top of that the file is reformatted and some minor changes are made
to the "build" job, namely:

- Add comments to visually separate CI stages
- Use version 2 of the _setup-go_ Action (which sets up environment
  variables for Go).
- Remove the step name when checking out the code