Proof of concept implementation for maintaining whitespace
This updates the implementation of the replacer with a proof of concept
implementation that maintains whitesapce (spaces, tabs, newlines) when
replacing words/phrases that contain spaces according to the source.

For example, if a mapping is specified as "hello world,hey planet" then
a source text with "hello\world" would previously be unchanged but now
will be updated to "hey\nplanet".

In addition, this covers edge cases. E.g. when the new line starts with
whitespace (tabs or spaces) then those are also maintained.

One side-effect, at least of the current implementation, is that a
mapping like "foo bar,foobar" will now replace both "foo bar" and
"foo  bar" (two spaces) with "foobar". This is not necessarily ideal...
(hence the change to `TestReplaceWhiteSpaceInPhrase`)