Update documentation, tests, and CHANGELOG for multi-whitespace mappings
Update the documentation to remove the note that multiple whitespace 
characters betwen words in a mapping phrase only match if they appear so 
in the input text, which is no longer the case.

Added a test case for a two-space `from` mapping value.

Also updated the CHANGELOG to include this (breaking) change.