Maintain capitalization in multiword values (#26)
* Maintain capitalization in phrases

This adds functionality that will maintain capitalization of all words 
in a multi-word from value (i.e. a phrase) under the assumption that the 
to value has the same number of words.

If len(from) != len(to) the capitalization behaviour is not clearly 
defined.

* Account for hyphens and other non-space characters that maintain caps

This improves the functionality that maintains capitalization for multi-
word values. Now, if the to value is split by e.g. a hyphen, instead of
a space, capitalization of the second word is still maintained.

Moreover, this fixes a potential bug where a multi-word value contains
multiple spaces between its words. The previous implementation assumed
a single space. The new implementation takes the division symbol(s)
found in the mapping.

* Update CHANGELOG

* Remove unused tests

* Refactor replacer package

1. Clean up the new `maintainCapitalization` function.

* Move formatting logic of replacer package to its own file

* Add documentation on capatalisation

Add documentation on how the program maintains capitalisation of mapping 
values. It covers maintaining the first capital letter, all caps, and 
capitalisation in multiple words.

* Test formatting module

Create a test suite for the formatting module of the replacer packages 
separate from the replacer module itself.

The testing uncovered a bug! Mapping from a, e.g., 2 word phrase to a 3 
word phrase will result in the only the first 2 words of the 3 word 
phrase being copied

* Fix big of missing words in replacement string

... when the `to` string has more words than the `from` string.