Replace phrases with spaces if source contains newlines (#37)
* Create test cases for phrases with spaces if the source has newlines

This adds test cases for the replacer packages for the scenario where
the "from" value contains a space (e.g. "a dog") and it appears in the
source text with a newline character in between (e.g. "a\ndog").

* Proof of concept implementation for maintaining whitespace

This updates the implementation of the replacer with a proof of concept
implementation that maintains whitesapce (spaces, tabs, newlines) when
replacing words/phrases that contain spaces according to the source.

For example, if a mapping is specified as "hello world,hey planet" then
a source text with "hello\world" would previously be unchanged but now
will be updated to "hey\nplanet".

In addition, this covers edge cases. E.g. when the new line starts with
whitespace (tabs or spaces) then those are also maintained.

One side-effect, at least of the current implementation, is that a
mapping like "foo bar,foobar" will now replace both "foo bar" and
"foo  bar" (two spaces) with "foobar". This is not necessarily ideal...
(hence the change to `TestReplaceWhiteSpaceInPhrase`)

* Update CHANGELOG

Add replacement of phrase matches with newlines to the list of new 
features.

* Add documentation on newline preserving functionality

* Update documentation, tests, and CHANGELOG for multi-whitespace mappings

Update the documentation to remove the note that multiple whitespace 
characters betwen words in a mapping phrase only match if they appear so 
in the input text, which is no longer the case.

Added a test case for a two-space `from` mapping value.

Also updated the CHANGELOG to include this (breaking) change.

* Create test suite for `maintainWhitespace` function