Initialize fuzzing support
This adds fuzzing support using go-fuzz [1]. Currently, there are three
fuzzing functions: one for the CLI arguments parser, one for parsing CSV
mapping files, and one for parsing MD mapping files.

In addition, there are now two fuzzing commands in the Makefile to make
running the fuzzer a little bit easier. To run the fuzzer for the CLI
package you can run `make fuzz-cli`. To rn the fuzzer for the wordmaps
package you can run `make fuzz-wordmaps FUNC=FuzzXXX`, where XXX should
be the name of the fuzz function for either CSV or MD files (`FuzzCsv`
and `fuzzMarkDown` respectively).

Lasty, the files generated by the fuzzer are always ignored, the clean
command will remove any files generated by the fuzer, and the fuzzer is
added as a (indirect) dependency.

--
1. https://github.com/dvyukov/go-fuzz