Fix crash due to empty argument
Add a fix to the CLI parsing logic that caused the program to crash on 
an empty argument.

Before the progam would always check if an argument starts with a dash 
by getting the first character as a string using `s[:1]`. But if the 
string has length 0 this panics.

This has been fixed by using `strings.HasPrefix` instead, which checks 
if a string starts with a certain prefix (string or character).