Generalize getParserForFormat to easily allow for multiple aliases
Update the `getParserForFormat` function to make it easy to add new 
aliases for formats in the future.

This is achieved by storing know aliases in a slice and having a 
`contains` function that checks if the specified format is in any of the 
slices for different formats.

This will help with the implementation of #44