Proof of concept of CLI specified mapping format separate from extension
Simple proof of concept (not-working) implementation that allows users 
to specify the format of a mapping file explicitly from the CLI, 
overriding the default format assumed from the file's extension.

The implementation is not working as the new function in the wordmaps 
package does not yet actually use the provided format. This currently 
depends on #45 being merged.