Refactor main
Extract logic to create a WordMap based on the input arguments from 
main.go and into a separate wordmap.go file in the main package. Also, 
the refactoring includes a small changes to the `getWordMap` function, 
simplying it just slightly.

To be able to do this the build target of the Makefile had to be 
updated.

In addition the unintentional change in wordmaps/csv.go is reverted.