Add support to explicitly state mapping format from CLI (#46)
* Proof of concept of CLI specified mapping format separate from extension

Simple proof of concept (not-working) implementation that allows users 
to specify the format of a mapping file explicitly from the CLI, 
overriding the default format assumed from the file's extension.

The implementation is not working as the new function in the wordmaps 
package does not yet actually use the provided format.

* Refactor main

Extract logic to create a WordMap based on the input arguments from 
main.go and into a separate wordmap.go file in the main package. Also, 
the refactoring includes a small changes to the `getWordMap` function, 
simplying it just slightly.

To be able to do this the build target of the Makefile had to be 
updated.

In addition the unintentional change in wordmaps/csv.go is reverted.

* Break `getWordMap` into two separate functions

Previously this function contained the code to process both map files 
and CLI specified mappings. These two functionalities are now exracted 
into helper functions called by `getWordMap`.

In addition, the logging in this package has been updated.

* Simplify usage of `parseWordMapArgument`

* Update CHANGELOG