Add build target to compile for both Windows and Linux
Add a build target (called "build-all") that compiles to program and
creates a binary for both Windows and Linux at once. The standard build
target is unchanged except for the fact that the binary file it
generates no longer has a file extensions.

Also update the clean target and .gitigore file accordingly