Update documentation of fuzzing
Update the documentation in the Contributing Guidelines on how to do
fuzzing for this project:

- Improve wording for clarity.
- Add template for fuzzing files.

The template for fuzzing files includes the `// +build gofuzz` comment
as the first line of the file, following [1]. This prevents the fuzzing
file from being included when the program is compiled (or when linting
tools are run, actually). The existing fuzzing files have been updated
accordingly.

--
1. https://github.com/dvyukov/go-fuzz#random-notes