Add functionality to escape and replace prefix and suffix hyphens
Add the ability to escape a prefix or suffix hyphen in a mapping to be
able to replace it. This now allows one to define the following mapping:

```csv
\-foo, foo
world\-, world!
```

Which can be used to replace "-foo" by foo and "world-" by "world!"
respectively.

This functionality is also tested.