Reconfigure Renovate and update Contributing Guidelines and the Makefile (#52)
* Update example git hook script in

Update the example git hook in the Contributing Guidelines to stash 
unstaged changes before testing and linting, inspired by [1].

--
1. 
https://codeinthehole.com/tips/tips-for-using-a-git-pre-commit-hook/#stash-unstaged-changes-before-running-tests

* Update Contributing Guidelines

- Add dividers between sections
- Mention the usage of gofmt [1]
- Rename "Golint" to "golint"
- Fix comment on go-fuzz prerequisite
- Add "Testing" section (content to be written), of which Fuzzing has
  become a subsection.
- Update the fuzzing shell comments for clarity.

--
1. https://golang.org/cmd/gofmt

* Update Makefile

- Remove unnecessary newline
- Remove unintended spaces in coverage command

* Update Renovate configuration

Move the Renovate configuration file to the .github folder. As the file 
is essentially useless outside GitHub it makes sense to put it there 
(given that Renovate allows you to [1]).

--
1. https://docs.renovatebot.com/configuration-options/

* Use consistent go version throughout project

Normalize the go version to 1.13 throughout the project. This means
updating the version in:

- go.mod
- CONTRIBUTING.md

* Add build target to compile for both Windows and Linux

Add a build target (called "build-all") that compiles to program and
creates a binary for both Windows and Linux at once. The standard build
target is unchanged except for the fact that the binary file it
generates no longer has a file extensions.

Also update the clean target and .gitigore file accordingly

* Update documentation of fuzzing

Update the documentation in the Contributing Guidelines on how to do
fuzzing for this project:

- Improve wording for clarity.
- Add template for fuzzing files.

The template for fuzzing files includes the `// +build gofuzz` comment
as the first line of the file, following [1]. This prevents the fuzzing
file from being included when the program is compiled (or when linting
tools are run, actually). The existing fuzzing files have been updated
accordingly.

--
1. https://github.com/dvyukov/go-fuzz#random-notes

* List available project commands in the Contributing Guidelines

Add a new section to the Contributing Guidelines which contains a table
of commands (`make xxx`) that are available in this project.

Also, the `make analyze` command has been renamed to `make analysis`.

* Draft content of Testing section in Contributing Guidelines

* Remove unnecessary quotes from `fuzz_dir` value in Makefile