Test replacing the escape character
With the ability of escaping the prefix/suffix hyphen using a backslash
the backslash is now also a special character. A test has been added to
the replacer's test suite to verify that backslashes are correctly
replaced.

This has lead to a small change in the implementation of the mappings
matcher, which no longer modifies the string based on its backslashes.