Add functionality to escape and replace hyphens in prefix and suffix position (#61)
* Add functionality to escape and replace prefix and suffix hyphens

Add the ability to escape a prefix or suffix hyphen in a mapping to be
able to replace it. This now allows one to define the following mapping:

```csv
\-foo, foo
world\-, world!
```

Which can be used to replace "-foo" by foo and "world-" by "world!"
respectively.

This functionality is also tested.

* Update CHANGELOG

* Test replacing the escape character

With the ability of escaping the prefix/suffix hyphen using a backslash
the backslash is now also a special character. A test has been added to
the replacer's test suite to verify that backslashes are correctly
replaced.

This has lead to a small change in the implementation of the mappings
matcher, which no longer modifies the string based on its backslashes.

* Document the prefix/suffix dash escaping mechanism