Add issue templates (#64)
* Add Bug report issue template

Add a GitHub issue template for bug reports with the information that is expected from a bug report.

Also updated the "Bug Reports" section in the Contributing Guidelines. Rewriting the entire section, inspired by [1], and linking to this template.

--
1. https://www.gnu.org/software/make/manual/html_node/Bugs.html#Bugs

* Add Feature request issue template

Add a GitHub issue template for feature requests with the information that is expected from a feature request.

Also updated the "Feature Requests" section in the Contributing Guidelines.

* Refine Bug Reports and Feature Requests sections

* Add contact links for GitHub Issues

Add a contact_link [1] linking to the Contributing Guidelines of the
project, inspired by [2].

--
1. https://help.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository#configuring-the-template-chooser
2. https://github.blog/2020-05-14-github-protips-tips-tricks-hacks-and-secrets-from-luke-hefson/