Improve the Flag type
Change the Flag type of the fs package from `uint16` to just `int` for
simplicity. Also, update the `OReadOnly` and `OReadWrite` values to be
of the type Flag. Lastly, added a `String()` method to the Flag type,
which is now used in replace.go. Lastly, the `ORead` flag has been
renamed to `OReadOnly`.