Update how the FUNC argument for fuzzing works
Update the working of the FUNC argument for the fuzz (and fuzz-build and
fuzz-run) targets. Before, it could be used optionally if a package has
multiple fuzzing targets, in which case you specify what comes after
"Fuzz". Now, if you specify FUNC you need to specify the whole function
name, including the Fuzz prefix. The default value is still just Fuzz.