Update the Contributing Guidelines
Update the description of the FUNC argument when running `make fuzz` to
reflect the fact that now the "Fuzz" prefix must be included.

In addition, the wording in the second paragraph has been adjusted.
"fuzzing code" has been replaced by "fuzzing logic" in accordance with
the comments in the provided Fuzz function template.