Update fuzz target (#81)
* Split the Makefile fuzz target into build and run

* Update how the FUNC argument for fuzzing works

Update the working of the FUNC argument for the fuzz (and fuzz-build and
fuzz-run) targets. Before, it could be used optionally if a package has
multiple fuzzing targets, in which case you specify what comes after
"Fuzz". Now, if you specify FUNC you need to specify the whole function
name, including the Fuzz prefix. The default value is still just Fuzz.

This change is based on the Makefile feature described in [1].

---
1. https://www.gnu.org/software/make/manual/make.html#Pattern_002dspecific


* Update the Contributing Guidelines

Update the description of the FUNC argument when running `make fuzz` to
reflect the fact that now the "Fuzz" prefix must be included.

In addition, the wording in the second paragraph has been adjusted.
"fuzzing code" has been replaced by "fuzzing logic" in accordance with
the comments in the provided Fuzz function template.