Fix WordMap.Iter test
The previous test assumed that the return value would be in a fixed
order. Now that it returns a `map[string]string` this is no longer the
case. Hence, the test has been updated to work correctly regardless of
the iteration order.

In addition, a note is added to the WordMap.Iter function documentation
to warn about this.