Move string matching logic to the replace package (#92)
* Move Mapping struct to replace

Properly separating concerns, and allow for a transition from the custom
WordMap type/struct to the replace package working with simple maps.

* Fix outdated docs in package replace & wordmaps

* Fix WordMap.Iter test

The previous test assumed that the return value would be in a fixed
order. Now that it returns a `map[string]string` this is no longer the
case. Hence, the test has been updated to work correctly regardless of
the iteration order.

In addition, a note is added to the WordMap.Iter function documentation
to warn about this.

* Remove old mapping.go file in package wordmaps

* Move Mapping struct to subpackage of replace

* Split Match struct into its own file

* Rename 'GetFrom' and 'GetTo' to 'From' and 'To'

* Add package documentation for new package mapping