Escape all special characters in getAllMatches
Update the getAllMatches function for the mapping package so that the
query strings (parameters `substr`) are escape. To achieve this the
QuoteMeta function from the regexp package [1] was used. This function
will automatically escape all special characters for regular epxression
in the provided string. Unfortuantely, this doesn't work very  nicely
with the mapping syntax prefix/suffix notation escape technique. For now
this is resolved by unescaping hyphens and backslashes before calling
QuoteMeta.

--
1. https://golang.org/pkg/regexp/#QuoteMeta