Add test to "replace" and change capitalization
This test tests the feature request of issue 87. This issue specifies
that it should be possible to fix capitalization of words/phrases using
wordrow. This test assumes the propsed implementation, where a mapping
that has two of the same values except for their capitalization shall be
interpreted as having to change the capitalization of words explicitly.