POC implementation of changing capitalisation
Proof of concept implementation of a replacement
where the capitalisation changes based on the mapping rather than the
document context. To be able to do this, the WordMap struct had to be
updated so that it no longer stores all values in all lowercase.
However, the replace package assumed this to be the case. Hence, this is
now implemented by providing the replacement in all-lowercase. Then, the
replaceOne function will check if this mapping is a "changes formatting"
mapping and ignore the formatting-maintained replacement and use the
exact To value instead.